Use the "nullptr" literal.
"nullptr" should be used to denote the null pointer cpp:S4962

Before C++11, the only way to refer to a null pointer was by using the integer literal 0, which created ambiguity about whether a pointer or an integer was intended. Even with the NULL macro, the underlying value is still 0.

C++11 introduced the keyword nullptr, which unambiguously refers to the null pointer. It should be used systematically.